refactor: Moved recipe input form to separate file
